Brain/requirements.txt
fastapi
uvicorn[standard]
jinja2
requests==2.26.0
python-dotenv==0.19.0
groq
openai
mangum
